add update handler

+++ internal/db/pg/pg.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/jmoiron/sqlx"
 	"github.com/kudras3r/EMobile/internal/config"
+	"github.com/kudras3r/EMobile/internal/db"
 	"github.com/kudras3r/EMobile/internal/models"
 	_ "github.com/lib/pq"
 )
@@ -29,6 +30,16 @@ func New(config config.DB) (*Storage, error) {
 	return &Storage{DB: db}, nil
 }
 
+func (s *Storage) songExists(id int) bool {
+	var exists bool
+	query := "SELECT EXISTS(SELECT 1 FROM songs WHERE id = $1)"
+	err := s.DB.QueryRow(query, id).Scan(&exists)
+	if err != nil {
+		return false
+	}
+	return exists
+}
+
 func (s *Storage) CloseConnection() {
 	s.DB.Close()
 }
@@ -39,6 +50,13 @@ func (s *Storage) GetConnection() *sql.DB {
 
 func (s *Storage) GetSongs(limit, offset int, filters *map[string]string) ([]models.Song, error) {
 	/* for paginating we use offset-based query */
+
+	if limit < 0 {
+		return nil, db.InvalidLimit(limit)
+	}
+	if offset < 0 {
+		return nil, db.InvalidOffset(offset)
+	}
 
 	songs := make([]models.Song, limit)
 	var b strings.Builder
@@ -81,6 +99,17 @@ func (s *Storage) GetSongs(limit, offset int, filters *map[string]string) ([]mod
 }
 
 func (s *Storage) GetSongText(id, limit, offset int) ([]models.Verse, error) {
+	if !s.songExists(id) {
+		return nil, db.SongNotExists(id)
+	}
+
+	if limit < 0 {
+		return nil, db.InvalidLimit(limit)
+	}
+	if offset < 0 {
+		return nil, db.InvalidOffset(offset)
+	}
+
 	verses := make([]models.Verse, limit)
 	query := fmt.Sprintf(
 		`SELECT * 
@@ -97,19 +126,36 @@ func (s *Storage) GetSongText(id, limit, offset int) ([]models.Verse, error) {
 	return verses, nil
 }
 
-func (s *Storage) DeleteSong(id int) (int, error) {
-	query := fmt.Sprintf("DELETE FROM songs WHERE id = %d", id)
-	res, err := s.DB.Exec(query)
+func (s *Storage) UpdateSong(id int, updated models.Song) (int, error) {
+	if !s.songExists(id) {
+		return -1, db.SongNotExists(id)
+	}
+
+	formatedDate := updated.ReleaseDate.Format("2006-01-02")
+	query := fmt.Sprintf(`UPDATE songs 
+						  SET title = '%s', group_performer = '%s', link = '%s', 
+						  release_date = '%s', verses_count = %d
+						  WHERE id = %d`,
+		updated.Song, updated.Group, updated.Link,
+		formatedDate, updated.VersesCount, id)
+
+	_, err := s.DB.Exec(query)
 	if err != nil {
 		return -1, err
 	}
 
-	rAffected, err := res.RowsAffected()
+	return id, nil
+}
+
+func (s *Storage) DeleteSong(id int) (int, error) {
+	if !s.songExists(id) {
+		return -1, db.SongNotExists(id)
+	}
+
+	query := fmt.Sprintf("DELETE FROM songs WHERE id = %d", id)
+	_, err := s.DB.Exec(query)
 	if err != nil {
 		return -1, err
-	}
-	if rAffected == 0 {
-		return -1, fmt.Errorf("no song found with id: %d", id)
 	}
 
 	return id, nil
